ref: separate index js to few modules
