+++ README.md
@@ -1,9 +1,11 @@
 # nodejs-file-manager
+
 File Manager uses Node.js APIs.
 
-The file manager is able to do the following:  
-- Work using CLI Perform basic file operations (copy, move, delete, rename, etc.) 
-- Utilize Streams API 
-- Get information about the host machine operating system 
-- Perform hash calculations 
+The file manager is able to do the following:
+
+- Work using CLI Perform basic file operations (copy, move, delete, rename, etc.)
+- Utilize Streams API
+- Get information about the host machine operating system
+- Perform hash calculations
 - Compress and decompress files
